+++ fridgeflix-frontend/.firebase/hosting.YnVpbGQ.cache
@@ -1,9 +1,9 @@
-asset-manifest.json,1696274897335,573ae7b3a3ceb5148415df6019ad9eeeccf1d3c7a7a501ccdfbfdc857c0b3d01
-index.html,1696274897335,1d34dbeae113226b332c3c56a614180263953296b5a1e283df3d4c477ba42ee1
-static/css/main.4066f7cf.css,1696274897341,3eeece4f3400e4402107b0eab7f38e32231626500450f0da7c0c9b60f5892d41
-static/css/main.4066f7cf.css.map,1696274897340,330b380bd5172429c671cbe2645b2b6d47ddb1d855c733bee39ecc270c26a9cb
-static/js/main.c923b12a.js.LICENSE.txt,1696274897340,51d661b9f06db1e544b52d2cf226d67371b12776c4ce8ee67adbd4fc5ddd635e
-robots.txt,1696274878933,391d14b3c2f8c9143a27a28c7399585142228d4d1bdbe2c87ac946de411fa9a2
-manifest.json,1696274878930,341d52628782f8ac9290bbfc43298afccb47b7cbfcee146ae30cf0f46bc30900
-static/js/main.c923b12a.js,1696274897340,d42721fc29f258bd8c87b7d7cc0a49a240c8353fdde8445306bb4d86959705c2
-static/js/main.c923b12a.js.map,1696274897341,7f7d08b47394b02ed8cecb2f6be3fe8b9093cbfc99c3de7bc837b63f93f555be
+asset-manifest.json,1696286081808,0d48636bc78ccb75f5d569431d5a698d0949f52d5fcfe1de53010003b3654381
+robots.txt,1696286051468,391d14b3c2f8c9143a27a28c7399585142228d4d1bdbe2c87ac946de411fa9a2
+manifest.json,1696286051466,341d52628782f8ac9290bbfc43298afccb47b7cbfcee146ae30cf0f46bc30900
+index.html,1696286081808,2f8238ced5aac54429f4d787e239b8f14b4d9c01afc1c457a526b5bd3c71462c
+static/js/main.5223ea86.js.LICENSE.txt,1696286081814,51d661b9f06db1e544b52d2cf226d67371b12776c4ce8ee67adbd4fc5ddd635e
+static/css/main.b6fc4e61.css,1696286081814,913ddd2a068e50cc46fdb27f43246dae60a8c227ee6572e78c19d3fe5136cadd
+static/css/main.b6fc4e61.css.map,1696286081814,c1890b3185aecd1df80563472edadd32e25c935352ded6ff6bb1dbbd97fe8cc7
+static/js/main.5223ea86.js,1696286081814,13ce1da3c8f0135dec878fadc614333c7dc9ab833606510dace2950597ed21f3
+static/js/main.5223ea86.js.map,1696286081814,96a23ae835fca0e0057430dd16bb67f3f8fcbf48a73a4dddbcc18557931f2af4
