+++ fridgeflix-frontend/src/Banner.jsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React from "react";
 
 const Banner = () => (
   <div className="banner">
