+++ fridgeflix-frontend/src/Constants.js
@@ -1,0 +1,67 @@
+export const VEGETABLES = [
+  "avocado",
+  "bell peppers",
+  "broccoli",
+  "cabbage",
+  "carrots",
+  "cauliflower",
+  "celery",
+  "cucumber",
+  "garlic",
+  "lettuce",
+  "mushrooms",
+  "olives",
+  "onions",
+  "peppers",
+  "potatoes",
+  "spinach",
+  "tomatoes",
+];
+
+export const PROTEINS = [
+  "anchovies",
+  "beef",
+  "bread",
+  "butter",
+  "cheese",
+  "chicken",
+  "clams",
+  "crab",
+  "duck",
+  "eggs",
+  "fish",
+  "lamb",
+  "lobster",
+  "salmon",
+  "sardines",
+  "shrimp",
+  "tofu",
+  "trout",
+  "turkey",
+  "tuna",
+];
+
+export const SPICES = [
+  "basil",
+  "black pepper",
+  "cardamom",
+  "chili powder",
+  "chives",
+  "cilantro",
+  "cinnamon",
+  "cloves",
+  "coriander",
+  "cumin",
+  "fennel",
+  "ginger",
+  "mint",
+  "nutmeg",
+  "oregano",
+  "paprika",
+  "parsley",
+  "rosemary",
+  "saffron",
+  "salt",
+  "thyme",
+  "turmeric",
+];
