+++ fridgeflix-frontend/src/Footer.jsx
@@ -1,0 +1,16 @@
+import React from 'react';
+
+const Footer = () => (
+    <div className="footer">
+      Created by{" "}
+      <a
+        href="https://www.linkedin.com/in/mazharalibaig/"
+        target="_blank"
+        rel="noopener noreferrer"
+      >
+        Mazhar Ali Baig
+      </a>
+    </div>
+  );
+
+export default Footer;
