--- fridgeflix-frontend/src/wordList.js
@@ -1,17 +1,0 @@
-export const WORD_CLOUD = [
-    "chicken",
-    "rice",
-    "potatoes",
-    "onions",
-    "bell peppers",
-    "beef",
-    "broccoli",
-    "bread",
-    "milk",
-    "eggs",
-    "cheese",
-    "lettuce",
-    "tomatoes",
-    "butter",
-    "garlic",
-  ];
